Extract sendfile tests into a separate test file
